+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- Render the empty tag expression as an empty string ([#222](https://github.com/cucumber/tag-expressions/pull/222))
+- Improve error message for missing operands ([#221](https://github.com/cucumber/tag-expressions/pull/221))
 
 - [dotNET] Add a .NET implementation